extract current column functions into their own selector file
