+++ test/unit/selectors.test.js
@@ -8,11 +8,7 @@ import {
   getSelectedNumericalFeatures,
   getUniqueOptionsByColumn,
   getUniqueOptionsLabelColumn,
-  getUniqueOptionsCurrentColumn,
-  getExtremaByColumn,
-  getExtremaCurrentColumn,
-  getOptionFrequenciesCurrentColumn,
-  isCurrentColumnReadOnly
+  getExtremaByColumn
 } from "../../src/selectors";
 import {
   classificationState,
@@ -74,23 +70,6 @@ describe("getting category options", () => {
     )
     expect(uniqueOptions).toEqual(['no', 'yes']);
   });
-
-  test("gets unique options current column", async () => {
-    const uniqueOptions = getUniqueOptionsCurrentColumn.resultFunc(
-      classificationState.currentColumn,
-      classificationState.data
-    )
-    expect(uniqueOptions).toEqual(['cool', 'hot', 'mild']);
-  });
-
-  test("gets option frequencies current column", async () => {
-    const optionFrequencies = getOptionFrequenciesCurrentColumn.resultFunc(
-      classificationState.currentColumn,
-      classificationState.data
-    )
-    expect(optionFrequencies).toEqual({ hot: 2, mild: 2, cool: 2 });
-  });
-
 });
 
 describe("getting extrema", () => {
@@ -116,33 +95,5 @@ describe("getting extrema", () => {
     expect(extremaByColumn['mosquitoCount'].max).toBe(10);
     expect(extremaByColumn['mosquitoCount'].min).toBe(1);
     expect(extremaByColumn['mosquitoCount'].range).toBe(9);
-  });
-
-  test("gets extrema current column", async () => {
-    const extrema = getExtremaCurrentColumn.resultFunc(
-      allNumericalState.currentColumn,
-      allNumericalState.data
-    )
-    expect(extrema.max).toBe(100);
-    expect(extrema.min).toBe(40);
-    expect(extrema.range).toBe(60);
-  });
-});
-
-describe("is current column readOnly", () => {
-  test("current column is readOnly", async () => {
-    const isReadOnly = isCurrentColumnReadOnly.resultFunc(
-      premadeDatasetState.currentColumn,
-      premadeDatasetState.metadata
-    )
-    expect(isReadOnly).toBe(true)
-  });
-
-  test("current column is not readOnly", async () => {
-    const isReadOnly = isCurrentColumnReadOnly.resultFunc(
-      allNumericalState.currentColumn,
-      allNumericalState.metadata
-    )
-    expect(isReadOnly).toBe(false)
   });
 });
